+++ src/promptflow-core/promptflow/core/_prompty_utils.py
@@ -3,9 +3,7 @@ import json
 import os
 import re
 from dataclasses import asdict
-from os import PathLike
-from pathlib import Path
-from typing import List, Mapping, Union
+from typing import List, Mapping
 
 from promptflow.core._connection import AzureOpenAIConnection, OpenAIConnection, _Connection
 from promptflow.core._errors import (
@@ -13,7 +11,6 @@ from promptflow.core._errors import (
     ChatAPIInvalidRoleError,
     CoreError,
     InvalidOutputKeyError,
-    InvalidSampleError,
     UnknownConnectionType,
 )
 from promptflow.core._model_configuration import ModelConfiguration
@@ -44,20 +41,6 @@ def parse_environment_variable(value):
         return os.environ.get(env_name, value)
     else:
         return value
-
-
-def load_inputs_from_sample(sample: Union[dict, str, PathLike]):
-    if not sample:
-        return {}
-    elif isinstance(sample, dict):
-        return sample
-    elif isinstance(sample, (str, Path)) and Path(sample).suffix.lower() == ".json":
-        if not Path(sample).exists():
-            raise InvalidSampleError(f"Cannot find sample file {sample}.")
-        with open(sample, "r") as f:
-            return json.load(f)
-    else:
-        raise InvalidSampleError("Only dict and json file are supported as sample in prompty.")
 
 
 def get_connection_by_name(connection_name):
