+++ src/promptflow-devkit/promptflow/_sdk/entities/_flows/flex.py
@@ -45,6 +45,10 @@ class FlexFlow(FlowBase):
     def additional_includes(self) -> list:
         return self._data.get("additional_includes", [])
 
+    @property
+    def sample(self):
+        return self._data.get("sample", None)
+
     # endregion
 
     # region overrides
